Give the survey-square tests files of their own, so the suite stops waiting on one (#419)
The JavaScript suite took sixty-three seconds on CI, and fifty of them
were four tests in stack-images-align.test.js: the ones that render each
scene at a phone's size and average it down through the pipeline's own
resize before the gate reads it. `node --test` runs files in parallel
and the tests inside a file one after another, so that file was the
critical path, with every other file finished in a quarter of the time
and waiting for it.

The scenes, squares and noise every alignment test draws move to
stack-images-fixtures.js, exported and asserting nothing, because a test
file cannot lend its fixtures out without also registering its tests in
whoever imports it. The four survey tests then split by which side of
the floor they pin: stack-images-survey-refused.test.js has the sky, the
sparse field and the floor-through-estimate, and
stack-images-survey-admitted.test.js the real families. Not one file for
the four, because the two longest tests are seventeen and sixteen
seconds and would still have been in series.

Nothing tested changes: the same thirty-six tests, the same fixtures
byte for byte, and locally the three files together run in twenty-five
seconds where the one file took thirty-five.
